pyproject.toml
[tool.poetry]
name = "corrective-rag"
version = "0.1.0"
description = ""
authors = ["Arman Tunga <armantunga@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
beautifulsoup4 = "^4.12.3"
langchain = "^0.3.0"
langgraph = "^0.2.23"
langchainhub = "^0.1.21"
langchain-community = "^0.3.0"
tavily-python = "^0.5.0"
langchain-qdrant = "^0.1.4"
python-dotenv = "^1.0.1"
ruff = "^0.6.7"
isort = "^5.13.2"
pytest = "^8.3.3"
langchain-openai = "^0.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
